adjust configuration, etc.
