+++ Marlin/src/gcode/bedlevel/abl/G29.cpp
@@ -36,7 +36,7 @@
 #include "../../../module/probe.h"
 #include "../../queue.h"
 
-#if ENABLED(PROBE_TEMP_COMPENSATION)
+#if HAS_PTC
   #include "../../../feature/probe_temp_comp.h"
   #include "../../../module/temperature.h"
 #endif
@@ -645,11 +645,9 @@ G29_TYPE GcodeSuite::G29() {
             break; // Breaks out of both loops
           }
 
-          #if ENABLED(PROBE_TEMP_COMPENSATION)
-            TERN_(USE_TEMP_BED_COMPENSATION, temp_comp.compensate_measurement(TSI_BED, thermalManager.degBed(), abl.measured_z));
-            TERN_(USE_TEMP_PROBE_COMPENSATION, temp_comp.compensate_measurement(TSI_PROBE, thermalManager.degProbe(), abl.measured_z));
-            TERN_(USE_TEMP_EXT_COMPENSATION, temp_comp.compensate_measurement(TSI_EXT, thermalManager.degHotend(0), abl.measured_z));
-          #endif
+          TERN_(USE_TEMP_BED_COMPENSATION,   temp_comp.compensate_measurement(TSI_BED,   thermalManager.degBed(),     abl.measured_z));
+          TERN_(USE_TEMP_PROBE_COMPENSATION, temp_comp.compensate_measurement(TSI_PROBE, thermalManager.degProbe(),   abl.measured_z));
+          TERN_(USE_TEMP_EXT_COMPENSATION,   temp_comp.compensate_measurement(TSI_EXT,   thermalManager.degHotend(0), abl.measured_z));
 
           #if ENABLED(AUTO_BED_LEVELING_LINEAR)
