+++ Marlin/src/inc/Conditionals_adv.h
@@ -550,6 +550,10 @@
   #endif
 #endif
 
+#if ANY(USE_TEMP_PROBE_COMPENSATION, USE_TEMP_BED_COMPENSATION, USE_TEMP_EXT_COMPENSATION)
+  #define HAS_PTC 1
+#endif
+
 // Let SD_FINISHED_RELEASECOMMAND stand in for SD_FINISHED_STEPPERRELEASE
 #if ENABLED(SD_FINISHED_STEPPERRELEASE)
   #ifndef SD_FINISHED_RELEASECOMMAND
