+++ platformio.ini
@@ -152,7 +152,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/gcode/calibrate/G33.cpp>
   -<src/gcode/calibrate/G34.cpp>
   -<src/gcode/calibrate/G34_M422.cpp>
-  -<src/gcode/calibrate/G76_M192_M871.cpp>
+  -<src/gcode/calibrate/G76_M871.cpp>
   -<src/gcode/calibrate/G425.cpp>
   -<src/gcode/calibrate/M12.cpp>
   -<src/gcode/calibrate/M48.cpp>
@@ -229,6 +229,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/gcode/sd/M808.cpp>
   -<src/gcode/temp/M104_M109.cpp>
   -<src/gcode/temp/M155.cpp>
+  -<src/gcode/temp/M192.cpp>
   -<src/gcode/units/G20_G21.cpp>
   -<src/gcode/units/M82_M83.cpp>
   -<src/gcode/units/M149.cpp>
